Create DeleteObjectScript.ts
This is inside a confirm button in a delete object dialog that pops up when a user presses delete for a object. The user enters a object id (could be changed into a more "friendly" identifier) in  a text entry box, then they presses delete and the dialog will show up.